synthetic_data_gen/requirements.txt
google-genai
vertexai
google-cloud-aiplatform[evaluation]
sqlglot
tqdm